Split class DiffusionCoeffForQuat
Put derivative computation in another class
